Move all notifications files into notifications feature
