+++ packages/client/src/features/notifications/components/index.ts
@@ -1,0 +1,3 @@
+import NotificationsContainer from "./NotificationsContainer";
+
+export { NotificationsContainer };
